+++ source/blender/nodes/shader/nodes/node_shader_bsdf_hair_principled.c
@@ -38,7 +38,7 @@ static bNodeSocketTemplate sh_node_bsdf_hair_principled_in[] = {
 	// Benedikt Bitterli's default brownish hair with PHEOmelanin
 	// The absorption coefficient below is 1.3f * pheomelanin mapping
 	{	SOCK_VECTOR,	1, N_("Absorption Coefficient"),		0.245531f, 0.52f, 1.365f, 0.0f, 0.0f, 1000.0f},
-	{	SOCK_FLOAT,		1, N_("Offset"),						2.f*M_PI/180.f, 0.0f, 0.0f, 0.0f, -M_PI_2, M_PI_2, PROP_ANGLE},
+	{	SOCK_FLOAT,		1, N_("Offset"),						2.f*((float)M_PI)/180.f, 0.0f, 0.0f, 0.0f, -M_PI_2, M_PI_2, PROP_ANGLE},
 	{	SOCK_FLOAT,		1, N_("RoughnessU"),					0.3f, 0.0f, 0.0f, 0.0f, 0.0f, 1.0f, PROP_FACTOR},
 	{	SOCK_FLOAT,		1, N_("RoughnessV"),					0.3f, 0.0f, 0.0f, 0.0f, 0.0f, 1.0f, PROP_FACTOR},
 	{	SOCK_FLOAT, 	1, N_("Roughness Randomization"), 		0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 1.0f, PROP_FACTOR},
